fix: model importing
